fixed #144. Возврат double из COM-объекта

+++ src/ScriptEngine/Machine/Contexts/COMWrapperContext.cs
@@ -105,7 +105,7 @@ namespace ScriptEngine.Machine.Contexts
             }
             else if (type == typeof(double))
             {
-                return ValueFactory.Create((decimal)objParam);
+                return ValueFactory.Create((decimal)(double)objParam);
             }
             else if (type == typeof(DateTime))
             {
